requirements.txt
discord.py
yt-dlp
ffmpeg
